feat: language toggle (#140)
<!-- Generated by sourcery-ai[bot]: start summary -->

## Summary by Sourcery

Add a LanguageToggle component to enable language switching in the
application. Enhance the ThemeToggle component and update navigation
items in the PublicNavigation component. Update language settings and
labels in the en-US.json file.

New Features:
- Introduce a LanguageToggle component to allow users to switch between
different languages in the application.

Enhancements:
- Update the ThemeToggle component to remove unnecessary className logic
for selected themes.
- Add new navigation items to the PublicNavigation component, including
links to SettleMint's main site and console.

Documentation:
- Update the en-US.json file to include new labels for navigation items,
such as 'About SettleMint' and 'Console'.

<!-- Generated by sourcery-ai[bot]: end summary -->