fix: hardhat deploy remote missing access token even after login, debug logs (#629)
1. `scs hardhat deploy remote --accept-defaults` (with defaults) was not
working after running `bunx settlemint login` (without defaults),
because after running `bunx settlemint login` without defaults a
sanitized value of `env.SETTLEMINT_INSTANCE` was used as the config key.
Then in `scs hardhat deploy remote --accept-defaults` a non-sanitized
value of `env.SETTLEMINT_INSTANCE` was being used to fetch the config.

In the IDEs, `env.SETTLEMINT_INSTANCE` =
`https://console.settlemint.com/` (with a trailing slash), so the config
was stored without the trailing slash in the key, was being fetched with
the trailing slash in the key.
    
Now we always return a sanitized value for the instance, so while
storing and fetching the same sanitized value is used.

2. Also adds `SETTLEMINT_DEBUG` env variable to show "debug" level logs.