feat(ci): Harden GitHub Actions runner and reduce permissions (#671)
This commit introduces two key changes to the GitHub Actions workflow:

1. Hardens the GitHub Actions runner by using the
`step-security/harden-runner` action. This
improves the security of the runner environment by enforcing an egress
policy.

2. Reduces the overall permissions granted to the workflow, limiting
them to the minimum
required set of `read-all` permissions. This follows the principle of
least privilege
   and reduces the potential attack surface.

## Summary by Sourcery

Restrict GitHub Actions permissions to read-all and harden the runner.

CI:
- Reduce workflow permissions to read-all.
- Harden the GitHub Actions runner with step-security/harden-runner and
enforce an audit egress policy.