feat: add scorecard and codeql workflows (#672)
This commit adds two new GitHub Actions workflows:

1. `scorecard.yml`: This workflow runs the OSSF Scorecard tool to
analyze the security and quality of the repository. It is scheduled to
run weekly on Saturdays.

2. `codeql.yml`: This workflow runs the CodeQL code analysis tool to
detect vulnerabilities in the JavaScript and TypeScript code. It is
triggered on push and pull request events to the main branch, as well as
a weekly schedule.

These workflows are important for ensuring the security and quality of
the project, and will help catch issues early in the development
process.

## Summary by Sourcery

Add CodeQL and Scorecard GitHub Actions workflows to improve code
security and quality. The CodeQL workflow analyzes JavaScript and
TypeScript code for vulnerabilities on push, pull request, and a weekly
schedule. The Scorecard workflow assesses the repository's security
posture weekly.

Enhancements:
- Move CodeQL and Scorecard workflows from the build workflow to
dedicated workflows.

CI:
- Add CodeQL workflow.
- Add Scorecard workflow.