feat: remove unused smart contract address env vars
The changes remove the `SETTLEMINT_SMART_CONTRACT_ADDRESS` and `SETTLEMINT_SMART_CONTRACT_DEPLOYMENT_ID` environment variables from the `DotEnvSchema` and the `writeEnvSpinner` function. This is because the smart contract address is now determined from the `hardhatConfig.networks?.btp?.from` value or the first available private key address, and is no longer stored in the environment variables.