feat: add default sender option to remote deployment (#708)
The changes in this commit add a new `--default-sender` option to the
remote deployment command in the Hardhat deployment script. This allows
users to specify a default sender address for the deployment, which can
be useful in scenarios where the deployment is being executed in a
non-interactive environment or where the user does not want to be
prompted for the sender address.

Additionally, the commit includes a change to the address prompt logic,
where the script will first check if a default sender is provided, and
if not, it will prompt the user for the address. This ensures that the
deployment can still be executed even if a default sender is not
provided.

## Summary by Sourcery

New Features:
- Added a `--default-sender` option to the `hardhat deploy remote`
command, allowing users to specify a default sender address, which
simplifies deployments in non-interactive environments.