fix: only select nodes with ecdsa or hsm key activated (#797)
## Summary by Sourcery

Fixes an issue where the CLI would attempt to deploy smart contracts to
nodes without an activated ECDSA or HSM key. It also improves error
handling during smart contract deployment and provides more informative
messages to the user.

Bug Fixes:
- Fixes an issue where the CLI would attempt to deploy smart contracts
to nodes without an activated ECDSA or HSM key.
- Improves error handling during smart contract deployment, providing
more informative messages to the user when a deployment is cancelled or
unsuccessful.